* 为 `RepoMgr` 添加单独元素的快捷Check，并添加可选的buff参数
* 在框架中添加行为树支持（尚未测试）

+++ Assets/GameLogic/Model/MonoManager/ConfigMgr/ModelConfig.cs
@@ -12,7 +12,7 @@ namespace GameLogic.Model.Mgr
 
 		public ulong DAY_TICKS;
 		public ulong NIGHT_TICKS;
-		public float VILL_ONE_MOVE_TICK;
+		public int VILL_ONE_MOVE_TICK;
 		public float NIGHT_TIME_SPEED;
 
 		[Header("村民体力系统配置")]
